Refactor model and dto paths
