+++ src/app/layout/new-event-type/new-event-type.component.ts
@@ -9,8 +9,8 @@ import { CommonModule } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { EventTypeService } from '../../services/event-type.service';
-import { EventType } from '../../model/event-type';
-import { CreateEventType } from '../../model/create-event-type';
+import { EventType } from '../../model/event/event-type';
+import { CreateEventType } from '../../services/dtos/event/create-event-type';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { ServiceService } from '../../services/service.service';
