+++ src/app/layout/new-product/new-product.component.ts
@@ -2,17 +2,17 @@ import { Component, ElementRef, ViewChild } from '@angular/core';
 import { FormArray, FormControl, FormGroup, FormsModule, Validators } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { MatSelectModule } from '@angular/material/select';
-import { Service } from '../../model/service';
+import { Service } from '../../model/service-product/service';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { EventTypeService } from '../../services/event-type.service';
-import { EventType } from '../../model/event-type';
+import { EventType } from '../../model/event/event-type';
 import { ServiceCategoryService } from '../../services/service-category.service';
-import { ServiceCategory } from '../../model/service-category';
+import { ServiceCategory } from '../../model/service-product/service-category';
 import { ProductService } from '../../services/product.service';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { Product } from '../../model/product';
+import { Product } from '../../model/service-product/product';
 import { take } from 'rxjs';
