--- src/app/model/ep-event.ts
@@ -1,4 +1,0 @@
-export class EPEvent {
-    constructor(public id: number = -1, public category: string = "Wedding", public name: string = '', public description: string = '') {
-    }
-}
