--- src/app/model/service-product-category.ts
@@ -1,4 +1,0 @@
-export interface ServiceProductCategory {
-    id: number;
-    name: string;
-}
