+++ src/app/parameters/home-service-product-filter-dialog-params.ts
@@ -1,4 +1,4 @@
-import { EventType } from "../model/event-type";
+import { EventType } from "../model/event/event-type";
 import { ServiceProductCategory } from "../model/service-product/service-product-category";
 import { ServiceProductFilteringValues } from "../services/dtos/service-product/service-product-filtering-values.dto";
 import { ServiceProductFilterParams } from "./service-product-filter-params";
