+++ src/app/services/product.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { Product } from '../model/product';
+import { Product } from '../model/service-product/product';
 
 @Injectable({
   providedIn: 'root'
