+++ src/app/services/profile.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, of, tap } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { User } from './dtos/user';
+import { User } from './dtos/user/user';
 
 @Injectable({
   providedIn: 'root'
