+++ src/app/services/service-category.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
-import { ServiceCategory } from '../model/service-category';
+import { ServiceCategory } from '../model/service-product/service-category';
 
 @Injectable({
   providedIn: 'root'
